src/backend/requirements.txt
numpy
opencv-python
Pillow
fastapi
python-multipart
uvicorn
